requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
Jinja2==3.1.2

python-dotenv==1.0.0

SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

celery==5.3.6
redis==5.0.1

pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1

jellyfish==1.0.3

requests==2.31.0

gunicorn==21.2.0
